+++ ClusterAlgorithm/Model/Atom.swift
@@ -10,9 +10,9 @@ import Foundation
 struct Atom: Hashable {
     var id: Int
     var type: Int
-    var x: Decimal
-    var y: Decimal
-    var z: Decimal
+    var x: Double
+    var y: Double
+    var z: Double
     var shouldBeConverted: Bool
     
     static let zero = Atom(id: 0,
